requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0

# Cache & Session
redis==5.0.1
aioredis==2.0.1

# AI & ML
openai==1.3.7
tiktoken==0.5.2
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8
pycryptodome==3.20.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1